[refactor] RTK 구조 변경으로 인한 변경 및 useHook 각 페이지에서 삭제
