+++ src/components/GetBlogAllComponent.tsx
@@ -2,8 +2,8 @@ import React, { useEffect, useCallback, useRef } from "react";
 import { useAppDispatch } from "../hooks";
 import { useSelector } from "react-redux";
 import { RootState } from "../store/store";
-import { getBlogAll } from "../slices/blogSlice";
-import { searchBlog } from "../slices/blogSlice";
+import { getBlogAll } from "../slices/blog/blogSlice";
+import { searchBlog } from "../slices/blog/blogSlice";
 import search from "../img/Search.png";
 import registerImg from "../img/register.png";
 import arrowLeft from "../img/arrowLeft.png";
