+++ src/components/LoginForm.tsx
@@ -21,7 +21,7 @@ import logo from "../img/logo.png";
 import circleFill from "../img/akar-icons_circle-check-fill.svg";
 import styled from "styled-components";
 import { useAppDispatch } from "../hooks";
-import { logIn } from "../slices/userSlice";
+import { logIn } from "../slices/user/userSlice";
 import { useSelector } from "react-redux";
 import { RootState } from "../store/store";
