+++ src/components/PostComment.tsx
@@ -1,46 +1,48 @@
 import React, { useCallback, useEffect, useState } from "react";
-import { GetBlogAllTr,PostCommentTable ,PostCommentInput,PostBlogBtn} from "./styled";
+import {
+  GetBlogAllTr,
+  PostCommentTable,
+  PostCommentInput,
+  PostBlogBtn,
+} from "./styled";
 import { useAppDispatch } from "../hooks";
 import { useSelector } from "react-redux";
 import { RootState } from "../store/store";
-import { registerComment } from "../slices/commentSlice";
+import { registerComment } from "../slices/comment/commentSlice";
 import { useLocation } from "react-router-dom";
 
 export default function PostComment() {
   const { comment, comments } = useSelector(
     (state: RootState) => state.comment
   );
-  const {user} = useSelector(
-    (state: RootState) => state.user
-  );
+  const { user } = useSelector((state: RootState) => state.user);
   const dispatch = useAppDispatch();
   const blogIdNum = Number(document.location.href.split("/:")[1]);
   console.log(document.location.href.split("/")[4]);
   const [commentData, setCommentData] = React.useState<postCommentType>({
-    blogId:blogIdNum,
-    request:{
-        content:""
-    }
+    blogId: blogIdNum,
+    request: {
+      content: "",
+    },
   });
   const commentHandler = (event: React.FormEvent<HTMLInputElement>) => {
     // console.log(event.currentTarget.value);
     setCommentData({
-        blogId:blogIdNum,
-        request:{
-            content: event.currentTarget.value
-        }
-    })
+      blogId: blogIdNum,
+      request: {
+        content: event.currentTarget.value,
+      },
+    });
   };
   const onSubmit = useCallback(
     (event: React.FormEvent<HTMLFormElement>) => {
       event.preventDefault();
       console.log(commentData);
-      if(user){
+      if (user) {
         dispatch(registerComment(commentData));
-      }else{
+      } else {
         alert("로그인을 선행해주세요!");
       }
-      
     },
     [dispatch, commentData]
   );
