+++ src/pages/IntroducePage.tsx
@@ -1,10 +1,8 @@
 import React from "react";
 import Header from "../subComponents/Header";
 import IntroduceComponent from "../components/IntroduceComponent";
-import useInquireMyInfo from "../useHooks/useInquireMyInfo";
 
 export default function IntroducePage() {
-  useInquireMyInfo();
   return (
     <div>
       <Header />
