+++ src/pages/MakeBlogPage.tsx
@@ -1,9 +1,7 @@
 import React from "react";
 import MakeBlogComponent from "../components/MakeBlogComponent";
-import useInquireMyInfo from "../useHooks/useInquireMyInfo";
 
 function MakeBlogPage() {
-  useInquireMyInfo();
   return <MakeBlogComponent />;
 }
