+++ src/pages/MakePortFolioSecondPage.tsx
@@ -1,10 +1,8 @@
 import React from "react";
 import MakePortfolioSecondComponent from "../components/MakePortfolioSecondComponent";
 import Footer from "../subComponents/Footer";
-import useInquireMyInfo from "../useHooks/useInquireMyInfo";
 
 function MakePortFolioSecondPage() {
-  useInquireMyInfo();
   return (
     <>
       <MakePortfolioSecondComponent />
