+++ src/pages/PortFolioPageSecond.tsx
@@ -2,10 +2,8 @@ import React from "react";
 import PortfolioPageSecondComponent from "../components/PortfolioPageSecondComponent";
 import PortfolioHeader from "../subComponents/PortfolioHeader";
 import Footer from "../subComponents/Footer";
-import useInquireMyInfo from "../useHooks/useInquireMyInfo";
 
 function PortFolioPageSecond() {
-  useInquireMyInfo();
   return (
     <div>
       <PortfolioHeader />
