+++ src/subComponents/PortfolioHeader.tsx
@@ -7,20 +7,16 @@ import {
   BlogHeaderProfile,
   LogoImg,
   HeaderLoginBtn,
-  HeaderLoginImg
+  HeaderLoginImg,
 } from "../components/styled";
 import { useNavigate } from "react-router-dom";
 import logo from "../img/logo.png";
-import { useAppDispatch } from "../hooks";
-import { getProfileImage } from "../slices/portFolioSlice";
 import { useAppSelector } from "../hooks";
 import { RootState } from "../store/store";
-import loginImg from "../img/login2.png"
+import loginImg from "../img/login2.png";
 function PortfolioHeader() {
-  
   const { user } = useAppSelector((state: RootState) => state.user);
   const navigate = useNavigate();
-  // const { user } = useSelector((state: RootState) => state.user);
   const userNickname = document.location.href.split("/:")[1];
   const navigateHome = () => {
     navigate(`/portfolio/get/:${userNickname}`);
@@ -28,7 +24,6 @@ function PortfolioHeader() {
 
   const navigatePortfolio = () => {
     navigate(`/portfolio/get/second/:${userNickname}`);
-    //portfolio 페이지로 이동 (기술스택 등등 보여주는 페이지)
   };
 
   const navigateBlog = () => {
@@ -39,22 +34,18 @@ function PortfolioHeader() {
     navigate("/");
   };
   const loginClick = () => {
-    navigate("/logIn")
-  }
-  const jwtToken = localStorage.getItem("jwtToken");
-  const dispatch = useAppDispatch();
-  useEffect(() => {
-    if (jwtToken) {
-      dispatch(getProfileImage());
-    }
-  }, [jwtToken]);
+    navigate("/logIn");
+  };
+
   return (
     <HeaderComponent>
       <BlogHeaderDetailWrapper>
         {user && user.profileImage ? (
           <BlogHeaderProfile src={user.profileImage}></BlogHeaderProfile>
         ) : (
-          <HeaderLoginBtn onClick={loginClick}><HeaderLoginImg src={loginImg}></HeaderLoginImg></HeaderLoginBtn>
+          <HeaderLoginBtn onClick={loginClick}>
+            <HeaderLoginImg src={loginImg}></HeaderLoginImg>
+          </HeaderLoginBtn>
         )}
 
         <div>{user && user.nickname ? user.nickname : ""}</div>
